The Dev and Pad1 item in Stat_t struct should be uint64 in 64-bits mips architecture.

+++ src/syscall/syscall_linux_mips64x.go
@@ -137,8 +137,8 @@ func Iopl(level int) (err error) {
 }
 
 type stat_t struct {
-	Dev        uint32
-	Pad0       [3]int32
+	Dev        uint64
+	Pad0       [3]int64
 	Ino        uint64
 	Mode       uint32
 	Nlink      uint32
